Redesigned, New Sounds, Many Flightcontroller
havy change to be compatible with many Fligthcontroller

new Soundfiles,

based on ae727f272b9d26e9c67a4ec7cd7d475ccb6c8972 2.005  all change after this are not included.

the next steps are, to test, and bring all new changes of the Main Branch to this Branch and then merge it to the Main.